projet_audit_app_mobile/requirements.txt
flask
python-dotenv
google-generativeai
werkzeug
pathlib
requests
